+++ ee/server/service/hostidentity/config.go
@@ -1,4 +1,4 @@
-package hostidscep
+package hostidentity
 
 import (
 	"context"
