+++ ee/server/service/hostidentity/scep.go
@@ -1,4 +1,4 @@
-package hostidscep
+package hostidentity
 
 import (
 	"context"
