+++ ee/server/service/hostidentity/types/host_identity_certificates.go
@@ -1,4 +1,4 @@
-package fleet
+package types
 
 import (
 	"crypto/ecdsa"
