+++ ee/server/service/hostidentity/types/host_identity_certificates_test.go
@@ -1,4 +1,4 @@
-package fleet
+package types
 
 import (
 	"crypto/ecdsa"
