+++ server/datastore/mysql/mysql.go
@@ -17,6 +17,7 @@ import (
 	"github.com/XSAM/otelsql"
 	"github.com/doug-martin/goqu/v9"
 	"github.com/doug-martin/goqu/v9/exp"
+	hostidscepdepot "github.com/fleetdm/fleet/v4/ee/server/service/hostidentity/depot"
 	"github.com/fleetdm/fleet/v4/server/config"
 	"github.com/fleetdm/fleet/v4/server/contexts/ctxdb"
 	"github.com/fleetdm/fleet/v4/server/contexts/ctxerr"
@@ -192,7 +193,7 @@ func (ds *Datastore) NewSCEPDepot() (scep_depot.Depot, error) {
 // NewHostIdentitySCEPDepot returns a scep_depot.Depot for host identity certs that uses the Datastore
 // underlying MySQL writer *sql.DB.
 func (ds *Datastore) NewHostIdentitySCEPDepot(logger log.Logger) (scep_depot.Depot, error) {
-	return newHostIdentitySCEPDepot(ds.primary, ds, logger)
+	return hostidscepdepot.NewHostIdentitySCEPDepot(ds.primary, ds, logger)
 }
 
 type entity struct {
